Se han corregido pequeños errores en las integraciones numéricas
Former-commit-id: ab7e7b84405a24d6916ee03b0870632b97104595

+++ src/Grid.f90
@@ -768,7 +768,7 @@ module Grid_
 		real(8), intent(in) :: x
 		integer :: output
 		
-		output = (x-this.min)/this.stepSize+1
+		output = nint((x-this.min)/this.stepSize)+1
 	end function pos
 	
 	!>
